packages/agent-runtime/pyproject.toml
[project]
name = "agent-runtime"
version = "0.1.0"
description = "Actor runtime system for agents"
requires-python = ">=3.8"
dependencies = [
    "pyzmq",
    "protobuf",
    "termcolor"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["agent_runtime"]

[tool.uv.sources]